[to be squashed] fix x_int64.c printing

+++ crypto/asn1/x_int64.c
@@ -10,6 +10,7 @@
 #include <stdio.h>
 #include "internal/cryptlib.h"
 #include "internal/asn1t.h"
+#include "internal/numbers.h"
 #include <openssl/bn.h>
 #include "asn1_locl.h"
 
@@ -78,8 +79,8 @@ static int uint64_print(BIO *out, ASN1_VALUE **pval, const ASN1_ITEM *it,
                         int indent, const ASN1_PCTX *pctx)
 {
     if ((it->size & INTxx_FLAG_SIGNED) == INTxx_FLAG_SIGNED)
-        return BIO_printf(out, "%jd\n", (intmax_t)*(int64_t *)pval);
-    return BIO_printf(out, "%ju\n", (uintmax_t)*(uint64_t *)pval);
+        return BIO_printf(out, "%jd\n", *(int64_t *)pval);
+    return BIO_printf(out, "%ju\n", *(uint64_t *)pval);
 }
 
 /* 32-bit variants */
@@ -145,8 +146,8 @@ static int uint32_print(BIO *out, ASN1_VALUE **pval, const ASN1_ITEM *it,
                         int indent, const ASN1_PCTX *pctx)
 {
     if ((it->size & INTxx_FLAG_SIGNED) == INTxx_FLAG_SIGNED)
-        return BIO_printf(out, "%ld\n", (intmax_t)*(int32_t *)pval);
-    return BIO_printf(out, "%lu\n", (uintmax_t)*(uint32_t *)pval);
+        return BIO_printf(out, "%d\n", *(int32_t *)pval);
+    return BIO_printf(out, "%u\n", *(uint32_t *)pval);
 }
